Fixed encoding processing problem in file creating and updating actions
In the previous implementation uses custom decoding method of
`lib.formatters.base64_decode` when `github.create_file` and
`github.update_file` actions are called with "encoding" parameter. And
those will be failed because of its implementation.

This removes `lib.formatters.base64_decode` method because it is not
used any more except for above actions, which is wrong the way of using.